% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_inheritance.R
\name{Blastoise}
\alias{Blastoise}
\title{Blastoise Class}
\description{
R6 class for a Blastoise.
}
\section{Class Constructor}{

\preformatted{Blastoise$new()}
}

\section{Public Methods}{

\describe{
    \item{\bold{\code{hydro_pump}}}{: The target is blasted by a huge volume
        of water launched under great pressure.
    }
}


\describe{
    \item{\bold{\code{water_gun}}}{: The target is blasted with a forceful
        shot of water.
    }
}


\describe{
    \item{\bold{\code{bubble}}}{: A spray of countless bubbles is jetted at
       the opposing Pokemon.
    }
}


\describe{
    \item{\bold{\code{bubble}}}{: A spray of countless bubbles is jetted at
       the opposing Pokemon.
    }
}
}

