% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pokemon_by_number}
\alias{get_pokemon_by_number}
\title{Pokedex Number Lookup}
\usage{
get_pokemon_by_number(pokedex_number)
}
\arguments{
\item{pokedex_number}{(whole number) Pokedex number of Pokemon to look up.}
}
\value{
Name of Pokemon as string
}
\description{
Given a Pokedex number, get name of corresponding Pokemon.
}
