% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{item_index}
\alias{item_index}
\title{xxx}
\format{
xxx
\describe{
\item{id}{xxx}
\item{item_name}{xxx}
\item{item_cost}{xxx}
\item{category}{xxx}
\item{attributes}{xxx}
\item{pokemon}{xxx}
\item{fling_power}{xxx}
\item{fling_effect}{xxx}
}
}
\source{
https://pokeapi.co/
}
\usage{
item_index
}
\description{
xxx
}
\keyword{datasets}
