% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokedex}
\alias{pokedex}
\title{National Pokédex}
\format{
A data frame of 24 rows representing Pokémon and 28 variables
\describe{
\item{natl_no}{Number in the National Pokédex.}
\item{name}{Name of Pokémon}
\item{type}{Type of Pokémon.}
\item{species}{Species of Pokémon.}
\item{height, weight}{Metric height and weight of Pokémon.}
\item{ability1, ability2, hidden_ability}{Abilities of Pokémon.}
\item{ev_yield}{Effort value yield of Pokémon.}
\item{catch_rate}{Catch rate of Pokémon.}
\item{base_friend}{Base friendship value of Pokémon.}
\item{base_exp}{Base experience value of Pokémon.}
\item{growth_rate}{Growth rate of Pokémon.}
\item{egg_group}{Egg group of Pokémon.}
\item{male}{Percentage of Pokémon that are male.}
\item{female}{Percentage of Pokémon that are female}
\item{egg_cycles}{Number of egg cycles for Pokémon.}
\item{base_hp}{Base health points of Pokémon.}
\item{base_attack}{Base attack of Pokémon.}
\item{base_defense}{Base defense of Pokémon.}
\item{base_spattack}{Base special attack of Pokémon.}
\item{base_spdefense}{Base special defense of Pokémon.}
\item{base_speed}{Base speed of Pokémon.}
\item{base_total}{Total value of Pokémon's base stats (excluding base experience and base friendship).}
\item{type_defenses}{Effectiveness of each type on Pokémon, in order: Normal, Fire, Water, Electric, Grass, Ice, Fighting, Posion, Ground, Flying, Psychic, Bug, Rock, Ghost, Dragon, Dark, Steel, and Fairy.}
\item{evo_order}{Evolution order of Pokémon.}
\item{evo_level}{Level that Pokémon evolves at.}
}
}
\source{
https://pokemondb.net/pokedex
}
\usage{
pokedex
}
\description{
A dataset comprising the National Pokédex from the Pokémon video games
}
\keyword{datasets}
