% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_article.R
\docType{package}
\name{pokemon-package}
\alias{pokemon}
\alias{pokemon-package}
\title{Pokemon: a Demo Package for roxygen2}
\description{
Demo package for "Level Up Your Package's Roxygen Documentation,"
   a talk about roxygen2 advanced tricks at satRday Chicago 2019.

   This article is package documentation and is accessible with
   \code{?pokemon} or with \code{?pokemon-package}.
}
\section{Check Out}{

   \url{https://chicago2019.satrdays.org/}
}

\author{
\strong{Maintainer}: Jay Qi \email{jayqi.opensource@gmail.com}

}
\keyword{internal}
