% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{Pokémon data}
\format{
A data frame with 949 rows and 22 columns:
\describe{
\item{id}{a numeric vector representing the ID of each Pokémon}
\item{pokemon}{a character vector representing the name of each Pokémon}
\item{species_id}{a numeric vector representing the species ID of each Pokémon}
\item{height}{a numeric vector representing the height of each Pokémon}
\item{weight}{a numeric vector representing the weight of each Pokémon}
\item{base_experience}{a numeric vector representing the base experience of each Pokémon}
\item{type_1}{a character vector representing the primary type of each Pokémon}
\item{type_2}{a character vector representing the secondary type of each Pokémon}
\item{hp}{a numeric vector representing the HP (hit points) of each Pokémon}
\item{attack}{a numeric vector representing the attack points of each Pokémon}
\item{defense}{a numeric vector representing the defense points of each Pokémon}
\item{special_attack}{a numeric vector representing the special attack points of each Pokémon}
\item{special_defense}{a numeric vector representing the special defense points of each Pokémon}
\item{speed}{a numeric vector representing the speed of each Pokémon}
\item{color_1}{a character vector representing the primary color of each Pokémon}
\item{color_2}{a character vector representing the secondary color of each Pokémon}
\item{color_f}{a character vector representing the final color of each Pokémon}
\item{egg_group_1}{a character vector representing the primary egg group of each Pokémon}
\item{egg_group_2}{a character vector representing the secondary egg group of each Pokémon}
\item{url_icon}{a character vector representing the URL image of each Pokémon}
\item{generation_id}{a numeric vector representing the generation ID of each Pokémon}
\item{url_image}{a character vector representing the URL image of each Pokémon}
}
}
\source{
\url{http://bulbapedia.bulbagarden.net}

\url{https://github.com/phalt/pokeapi}

\url{http://pokemon-uranium.wikia.com}

\url{https://github.com/HybridShivam/Pokemon}
}
\usage{
pokemon
}
\description{
A dataset with pokémon data from Nintendo games.
}
\keyword{datasets}
