% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokemon_ptbr}
\alias{pokemon_ptbr}
\title{Pokémon data (PT-BR)}
\format{
Um data frame com 949 linhas e 22 colunas:
\describe{
\item{id}{um vetor numérico representando o ID de cada Pokémon}
\item{pokemon}{um vetor numérico representando a altura de cada Pokémon}
\item{id_especie}{um vetor de caracteres representando o nome de cada Pokémon}
\item{altura}{um vetor numérico representando o peso de cada Pokémon}
\item{peso}{um vetor numérico representando a experiência base de cada Pokémon}
\item{exp_base}{um vetor de caracteres representando o tipo primário de cada Pokémon}
\item{tipo_1}{um vetor de caracteres representando o tipo secundário de cada Pokémon}
\item{tipo_2}{um vetor numérico representando os pontos de vida (HP) de cada Pokémon}
\item{hp}{um vetor numérico representando os pontos de ataque de cada Pokémon}
\item{ataque}{um vetor numérico representando os pontos de defesa de cada Pokémon}
\item{defesa}{um vetor numérico representando os pontos de ataque especial de cada Pokémon}
\item{ataque_especial}{um vetor numérico representando os pontos de defesa especial de cada Pokémon}
\item{defesa_especial}{um vetor numérico representando a velocidade de cada Pokémon}
\item{velocidade}{um vetor de caracteres representando a cor primária de cada Pokémon}
\item{cor_1}{um vetor de caracteres representando a cor secundária de cada Pokémon}
\item{cor_2}{um vetor de caracteres representando a cor final de cada Pokémon}
\item{cor_final}{um vetor de caracteres representando o grupo de ovos primário de cada Pokémon}
\item{grupo_ovo_1}{um vetor de caracteres representando o grupo de ovos primário de cada Pokémon}
\item{grupo_ovo_2}{um vetor de caracteres representando o grupo de ovos secundário de cada Pokémon}
\item{url_icone}{um vetor de caracteres representando a URL da imagem de ícone de cada Pokémon}
\item{url_imagem}{um vetor de caracteres representando o grupo de ovos secundário de cada Pokémon}
\item{id_geracao}{um vetor numérico representando o ID de geração de cada Pokémon}
}
}
\source{
\url{https://bulbapedia.bulbagarden.net/wiki/Main_Page}

\url{https://github.com/phalt/pokeapi}

\url{https://pokemon-uranium.wikia.com}

\url{https://github.com/HybridShivam/Pokemon}
}
\usage{
pokemon_ptbr
}
\description{
Uma base de dados de Pokémon, extraída dos jogos da Nintendo.
}
\keyword{datasets}
