% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Squirtle}
\alias{Squirtle}
\title{Squirtle Class}
\description{
R6 class for a Squirtle Pokemon.
}
\section{Class Constructor}{

\preformatted{Squirtle$new()}
}

\section{Public Methods}{

\describe{
    \item{\bold{\code{tackle}}}{: A physical attack in which the user charges
        and slams into the target with its whole body.
    }
    \item{\bold{\code{tail_whip}}}{: The user wags its tail cutely, making
        opposing Pokemon less wary and lowering their Defense stat.
    }
}
}

