% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_to_pokemon}
\alias{translate_to_pokemon}
\alias{translate_to_pikachu}
\alias{translate_to_squirtle}
\title{Translate text to Pokemon speech}
\usage{
translate_to_pikachu(sentence)

translate_to_squirtle(sentence)
}
\arguments{
\item{sentence}{(character string) text to translate}
}
\value{
Translated string
}
\description{
Given a sentence of text, translate it to a particular Pokemon's
   speech.
}
\examples{
translate_to_pikachu("I want to be the very best,")
translate_to_squirtle("Like no one ever was.")
}
