% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Wartortle}
\alias{Wartortle}
\title{Wartortle Class}
\description{
R6 class for a Wartortle Pokemon.
}
\section{Class Constructor}{

\preformatted{Wartortle$new()}
}

\section{Public Methods}{

\describe{
    \item{\bold{\code{water_gun}}}{: The target is blasted with a forceful
        shot of water.
    }
}


\describe{
    \item{\bold{\code{tackle}}}{: A physical attack in which the user charges
        and slams into the target with its whole body.
    }
    \item{\bold{\code{tail_whip}}}{: The user wags its tail cutely, making
        opposing Pokemon less wary and lowering their Defense stat.
    }
}
}

