% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_inheritance.R
\name{whos_that_pokemon}
\alias{whos_that_pokemon}
\title{Who's that Pokemon?}
\usage{
whos_that_pokemon(pokedex_number, print_result = TRUE)
}
\arguments{
\item{pokedex_number}{(whole number) Pokedex number of Pokemon to look up.}

\item{print_result}{(logical) whether to print result with \code{cat}}
}
\value{
string
}
\description{
"Who's that Pokemon?" by Pokedex number.
}
