% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Blastoise}
\alias{Blastoise}
\title{Blastoise Class}
\description{
R6 class for a Blastoise Pokemon.
}
\section{Class Constructor}{

\preformatted{Blastoise$new()}
}

\section{Public Methods}{

\describe{
    \item{\bold{\code{hydro_pump}}}{: The target is blasted by a huge volume
        of water launched under great pressure.
    }
}


\describe{
    \item{\bold{\code{water_gun}}}{: The target is blasted with a forceful
        shot of water.
    }
}


\describe{
    \item{\bold{\code{tackle}}}{: A physical attack in which the user charges
        and slams into the target with its whole body.
    }
    \item{\bold{\code{tail_whip}}}{: The user wags its tail cutely, making
        opposing Pokemon less wary and lowering their Defense stat.
    }
}


\describe{
    \item{\bold{\code{tackle}}}{: A physical attack in which the user charges
        and slams into the target with its whole body.
    }
    \item{\bold{\code{tail_whip}}}{: The user wags its tail cutely, making
        opposing Pokemon less wary and lowering their Defense stat.
    }
}
}

