% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{item_index}
\alias{item_index}
\title{Available Items}
\format{
A data frame of 1607 rows and 7 variables:
\describe{
\item{item_name}{Name of item.}
\item{item_cost}{Price of item in stores within game.}
\item{category}{Category of item falls into.}
\item{attributes}{List of attributes, or characterists, about item.}
\item{pokemon}{List of Pokémon that can hold/use item.}
\item{fling_power}{Power of using move "Fling" with this item.}
\item{fling_effect}{Effect of using move "Fling" with this item.}
}
}
\source{
https://pokeapi.co/item
}
\usage{
item_index
}
\description{
A data frame representing items available in Pokémon games. Items can be picked up and carried by both Pokémon and their trainers, and have a variety of uses.
}
\keyword{datasets}
