% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{item_index}
\alias{item_index}
\title{Available Items}
\format{
A data frame of 1607 rows and 7 variables:
\describe{
\item{item_name}{Name of the item.}
\item{item_cost}{Price of the item in stores within the game.}
\item{category}{Category of items that this item falls into.}
\item{attributes}{A list of attributes, or characterists, about the item}
\item{pokemon}{A list of Pokémon that can hold/use the item.}
\item{fling_power}{Power of using the move "Fling" with this item.}
\item{fling_effect}{Effect of using the move "Fling" with this item.}
}
}
\source{
https://pokeapi.co/
}
\usage{
item_index
}
\description{
A data frame representing items available in Pokémon games. Items can be picked up and carried by both Pokémon and their trainers, and have a variety of uses.
}
\keyword{datasets}
