% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moves_index}
\alias{moves_index}
\title{Pokémon moves}
\format{
A data frame of 844 rows representing Pokémon moves and 8 variables
\describe{
\item{move_name}{Name of move.}
\item{type}{The elemental type, or category, this move belongs in, e.g., ice.}
\item{accuracy}{Percent value of how likely this move is to be successful.}
\item{pp}{Power Points - The number of times this move can be used.}
\item{damage_class}{The type of damage the move inflicts on the target, e.g. physical.}
\item{power}{The base power of this move with a value of 0 if it does not have a base power.}
\item{learned_by_pokemon}{List of Pokemon that can learn the move.}
\item{generation}{The generation in which this move was introduced.}
}
}
\source{
https://pokeapi.co/
}
\usage{
moves_index
}
\description{
A collection of data about moves and associated information in the Pokémon games.
}
\keyword{datasets}
