% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moves_index}
\alias{moves_index}
\title{xxx}
\format{
xxx
\describe{
\item{move_name}{xxx}
\item{type}{xxx}
\item{accuracy}{xxx}
\item{pp}{xxx}
\item{damage_class}{xxx}
\item{power}{xxx}
\item{learned_by_pokemon}{xxx}
\item{generation}{xxx}
}
}
\source{
https://pokeapi.co/
}
\usage{
moves_index
}
\description{
xxx
}
\keyword{datasets}
