% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokedex}
\alias{pokedex}
\title{Pokedex Dataset (Gen 1)}
\format{data.frame}
\source{
\url{https://www.kaggle.com/rounakbanik/pokemon/version/1}
}
\usage{
pokedex
}
\description{
Pokedex data for generation 1 Pokemon.
}
\keyword{datasets}
