% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\docType{package}
\name{pokemon-package}
\alias{pokemon}
\alias{pokemon-package}
\title{Pokemon: a Demo Package for roxygen2}
\description{
Demo package for "Level Up Your Package's Roxygen Documentation,"
   a talk about roxygen2 advanced tricks at satRday Chicago 2019.

   This article is package documentation and is accesible with \code{?pokemon}
   or with \code{?pokemon-package}.
}
\section{Other stuff}{


Check out:
\itemize{
   \item{\code{\link{whos_that_pokemon}}}
}
}

\author{
\strong{Maintainer}: Jay Qi \email{jayqi.opensource@gmail.com}

}
\keyword{internal}
