% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_inheritance.R
\name{Squirtle}
\alias{Squirtle}
\title{Squirtle Class}
\description{
R6 class for a Squirtle.
}
\section{Class Constructor}{

\preformatted{Squirtle$new()}
}

\section{Public Methods}{

\describe{
    \item{\bold{\code{bubble}}}{: A spray of countless bubbles is jetted at
       the opposing Pokemon.
    }
}
}

