% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_inheritance.R
\name{Wartortle}
\alias{Wartortle}
\title{Wartortle Class}
\description{
R6 class for a Wartortle.
}
\section{Class Constructor}{

\preformatted{Wartortle$new()}
}

\section{Public Methods}{

\describe{
    \item{\bold{\code{water_gun}}}{: The target is blasted with a forceful
        shot of water.
    }
}


\describe{
    \item{\bold{\code{bubble}}}{: A spray of countless bubbles is jetted at
       the opposing Pokemon.
    }
}
}

